feat: add InventoryStartedFixedItems component
